generate TOC

+++ src/main/scala/dog/autodoc/AutodocPlugin.scala
@@ -11,22 +11,34 @@ object AutodocPlugin extends AutoPlugin {
     val autodocClean = TaskKey[Unit]("autodocClean")
     val autodocEnable = SettingKey[Boolean]("autodocEnable")
     val autodocTrimNameRegex = SettingKey[String]("autodocTrimNameRegex")
+    val autodocToc = SettingKey[Boolean]("autodocToc")
+    val autodocInitializeToc = TaskKey[Unit]("autodocInitializeToc")
 
     object Default {
       val outputDir = "doc"
       val enable = false
       val trim = "(Test|Spec)$"
+      val toc = false
     }
 
     val autodocSettings: Seq[Setting[_]] = Seq(
       testFrameworks += new TestFramework("dog.autodoc.AutodocFramework"),
       autodocOutputDirectory := Default.outputDir,
       autodocClean := {
-        Task.deleteFile(baseDirectory.value, autodocOutputDirectory.value)
+        if(autodocEnable.value) Task.deleteFile(baseDirectory.value, autodocOutputDirectory.value)
       },
-      test <<= (test in Test) dependsOn (autodocClean),
+      autodocInitializeToc := {
+        if(autodocEnable.value && autodocToc.value) {
+          val outputDir = file(autodocOutputDirectory.value)
+          if(!outputDir.exists()) IO.createDirectory(outputDir)
+          val toc = outputDir / Markdown.tocFileName
+          IO.write(toc, Markdown.tocTitle)
+        }
+      },
+      test <<= (test in Test) dependsOn (autodocClean, autodocInitializeToc),
       autodocEnable := Default.enable,
       autodocTrimNameRegex := Default.trim,
+      autodocToc := Default.toc,
       testListeners ++= {
         if(autodocEnable.value) Seq(
           new dog.autodoc.AutodocListener(file(autodocOutputDirectory.value), autodocTrimNameRegex.value)
